feat(nodes): organise/tidy
